+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+#### A.1.1
+ - Adjusted cars position
+ - Added sterring support for touch devices
+ - Removed achievements code
+
 #### A.1.0
  - Fork carride to it's own project